Move metrics tracker from plugin reporting to shared utils
Relocated createMetricsTracker from src/plugin/src/reporting/metrics-tracker.js to src/shared/utils/metrics.js so metrics code used by identifier-case and project index can live in a neutral shared module. Updated imports, barrels, and tests to point at the shared entry point.
